test: remove unused code
