--- e2e/doc/condition_table.md
@@ -1,17 +1,0 @@
- ## node_power_off.robot
-  test case | node 1 | node 2 | node 3
- --|--|--|--
- 1 | <br><li>replica</li><li>attached</li><li>power off</li><br> | replica | replica
- 2 | <br><li>replica</li><li>attached</li><br> | <br><li>replica</li><li>power off</li><br>  | replica
- 3 | <br><li>replica</li><br> | <br><li>replica</li><br>  | <br><li>attached</li><li>power off</li><br>
- 4 | <br><li>replica</li><li>power off</li><br> | <br><li>replica</li><br>  | <br><li>attached</li><br>
- 5 | <br><li>replica</li><li>attached</li><li>power off</li><br> | <br><li>replica</li><br>  |
- 6 | <br><li>replica</li><li>attached</li><br> | <br><li>replica</li><br>  |<br><li>power off</li><br>
- 7 | <br><li>replica</li><li>attached</li><br> | <br><li>replica</li><li>power off</li><br>  |
- 8 | <br><li>replica</li><li>attached</li><li>power off time out</li><br> | replica | replica
- 9 | <br><li>replica</li><li>attached</li><br> | <br><li>replica</li><li>power off time out</li><br>  | replica
- 10 | <br><li>replica</li><br> | <br><li>replica</li><br>  | <br><li>attached</li><li>power off time out</li><br>
- 11 | <br><li>replica</li><li>power off  time out</li><br> | <br><li>replica</li><br>  | <br><li>attached</li><br>
- 12 | <br><li>replica</li><li>attached</li><li>power off time out</li><br> | <br><li>replica</li><br>  |
- 13 | <br><li>replica</li><li>attached</li><br> | <br><li>replica</li><br>  |<br><li>power off time out</li><br>
- 14 | <br><li>replica</li><li>attached</li><br> | <br><li>replica</li><li>power off time out</li><br>  |
