+++ e2e/libs/node/node.py
@@ -68,3 +68,29 @@ class Node:
         waiter = self.aws_client.get_waiter('instance_running')
         waiter.wait(InstanceIds=instance_ids)
         logging(f"Started instances")
+
+    def get_all_pods_on_node(self, node_name):
+        api = client.CoreV1Api()
+        all_pods = api.list_namespaced_pod(namespace='longhorn-system', field_selector='spec.nodeName=' + node_name)
+        user_pods = [p for p in all_pods.items if (p.metadata.namespace != 'kube-system')]
+        return user_pods
+
+    def wait_all_pods_evicted(self, node_name):
+        for i in range(RETRY_COUNT):
+            pods = self.get_all_pods_on_node(node_name)
+            evicted = True
+            for pod in pods:
+                # check non DaemonSet Pods are evicted or terminating (deletionTimestamp != None)
+                pod_type = pod.metadata.owner_references[0].kind
+                pod_delete_timestamp = pod.metadata.deletion_timestamp
+
+                if pod_type != 'DaemonSet' and pod_delete_timestamp == None:
+                    evicted = False
+                    break
+
+            if evicted:
+                break
+
+            time.sleep(RETRY_INTERVAL)
+
+        assert evicted, 'failed to evict pods'
