--- e2e/libs/utils/__init__.py
@@ -1,2 +1,0 @@
-from utils import config_utils
-from utils import common_utils
