--- e2e/settings.ini
@@ -1,4 +1,0 @@
-[DEFAULT]
-CLOUD_PROVIDER = aws
-K8S_DISTRO = rke2
-LONGHORN_CLIENT_URL = http://127.0.0.1:8080/
